@@ -8,7 +8,7 @@ import setuptools
 
 
 MODULE_NAME = "sqlitebiter"
-REPOSITORY_URL = "https://github.com/thombashi/{:s}".format(MODULE_NAME)
+REPOSITORY_URL = f"https://github.com/thombashi/{MODULE_NAME:s}"
 REQUIREMENT_DIR = "requirements"
 ENCODING = "utf8"
 
@@ -68,13 +68,13 @@ setuptools.setup(
     long_description_content_type="text/x-rst",
     packages=setuptools.find_packages(exclude=["test*"]),
     project_urls={
-        "Documentation": "https://{:s}.rtfd.io/".format(MODULE_NAME),
-        "Download": "{:s}/releases".format(REPOSITORY_URL),
+        "Documentation": f"https://{MODULE_NAME:s}.rtfd.io/",
+        "Download": f"{REPOSITORY_URL:s}/releases",
         "Source": REPOSITORY_URL,
-        "Tracker": "{:s}/issues".format(REPOSITORY_URL),
-        "Changes": "{:s}/releases".format(REPOSITORY_URL),
+        "Tracker": f"{REPOSITORY_URL:s}/issues",
+        "Changes": f"{REPOSITORY_URL:s}/releases",
     },
-    python_requires=">=3.5",
+    python_requires=">=3.6",
     install_requires=install_requires,
     extras_require={
         "all": gs_requires + mediawiki_requires + optional_requires,
@@ -95,7 +95,6 @@ setuptools.setup(
         "Operating System :: POSIX",
         "Operating System :: POSIX :: Linux",
         "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.5",
         "Programming Language :: Python :: 3.6",
         "Programming Language :: Python :: 3.7",
         "Programming Language :: Python :: 3.8",
