@@ -55,7 +55,7 @@ def main():
         PROJECT_NAME,
         OUTPUT_DIR,
         is_make_toc=True,
-        project_url="https://github.com/thombashi/{}".format(PROJECT_NAME),
+        project_url=f"https://github.com/thombashi/{PROJECT_NAME}",
     )
     maker.examples_dir_name = "usage"
 
@@ -70,7 +70,7 @@ def main():
 
     maker.set_indent_level(0)
     maker.write_chapter("Documentation")
-    maker.write_lines(["https://{:s}.rtfd.io/".format(PROJECT_NAME.lower())])
+    maker.write_lines([f"https://{PROJECT_NAME.lower():s}.rtfd.io/"])
 
     maker.write_file(maker.doc_page_root_dir_path.joinpath("sponsors.rst"))
 
