@@ -1,5 +1,5 @@
 __author__ = "Tsuyoshi Hombashi"
-__copyright__ = "Copyright 2016, {}".format(__author__)
+__copyright__ = f"Copyright 2016, {__author__}"
 __license__ = "MIT License"
 __version__ = "0.34.1"
 __maintainer__ = __author__
