@@ -7,11 +7,11 @@ import traceback
 
 
 def print_test_result(expected, actual):
-    print("[expected]\n{}\n".format(expected))
-    print("[actual]\n{}\n".format(actual))
+    print(f"[expected]\n{expected}\n")
+    print(f"[actual]\n{actual}\n")
 
 
 def print_traceback(result):
     traceback.print_tb(result.exc_info[2], file=sys.stdout)
-    print("{}\n{}\n".format(result.exc_info[0], result.exc_info[1]))
-    print("[output]\n{}".format(result.output))
+    print(f"{result.exc_info[0]}\n{result.exc_info[1]}\n")
+    print(f"[output]\n{result.output}")
